requirements.txt
slack-bolt==1.18.0
spacy==3.7.2
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
streamlit==1.28.0
plotly==5.18.0
pandas==2.0.3
plotly==5.18.0
streamlit-autorefresh==1.0.1
slack-sdk==3.23.0 
transformers[torch]==4.36.2
prophet==1.1.5
accelerate==0.27.2
tabulate==0.9.0
flask-limiter 
python-dotenv-vault
huggingface_hub[hf_xet]
python-dateutil